💫 Add new file for data reading
