+++ gender_energy_networks/__init__.py
@@ -3,3 +3,4 @@
 __version__ = "0.1.0"
 from .enigh_class import *  # noqa
 from .data_processing_functions import *  # noqa
+from .get_dataframes_from_raw_data import *  # noqa
